REFACTOR component structure: Put components into folders
