+++ frontend/src/Components/App/App.test.tsx
@@ -1,5 +1,5 @@
 import { act, fireEvent, render } from '@testing-library/preact';
-import { SCALES } from '../constants';
+import { SCALES } from '../../constants';
 import { App } from './App';
 
 const ConfigureMockWebSocket = () => {
