+++ frontend/src/Components/App/App.tsx
@@ -1,7 +1,7 @@
 import classes from './App.module.css';
-import { PageSelector } from './PageSelector';
-import { TouchDetector } from './TouchDetector';
-import { WebSocketProvider } from './WebSocket';
+import { PageSelector } from '../PageSelector/PageSelector';
+import { TouchDetector } from '../TouchDetector/TouchDetector';
+import { WebSocketProvider } from '../WebSocket/WebSocket';
 
 export const App = () => (
   <div class={classes.app}>
