+++ frontend/src/Components/PageSelector/PageSelector.module.css
