+++ frontend/src/Components/RevealButton/RevealButton.module.css
@@ -1,5 +1,5 @@
 .revealButton {
-  composes: button from '../styles.module.css';
+  composes: button from '../../styles.module.css';
   margin-top: 2.5rem;
   margin-bottom: 0.5rem;
   height: 4.5rem;
