+++ frontend/src/Components/RevealButton/RevealButton.tsx
@@ -1,7 +1,7 @@
-import { Votes } from '../types/WebSocket';
+import { Votes } from '../../types/WebSocket';
 import classes from './RevealButton.module.css';
-import { connectToWebSocket } from './WebSocket';
-import { BUTTON_REVEAL_NOW, BUTTON_REVEAL_VOTES, VOTE_NOTE_VOTED } from '../constants';
+import { connectToWebSocket } from '../WebSocket/WebSocket';
+import { BUTTON_REVEAL_NOW, BUTTON_REVEAL_VOTES, VOTE_NOTE_VOTED } from '../../constants';
 
 const getNumberOfMissingVotes = (votes: Votes): number =>
   Object.values(votes).reduce((count, vote) => (vote === VOTE_NOTE_VOTED ? count + 1 : count), 0);
