+++ frontend/src/Components/TouchDetector/TouchDetector.tsx
