+++ frontend/src/Components/VotingStateDisplay/VotingStateDisplay.module.css
