+++ frontend/src/Components/VotingStateDisplay/VotingStateDisplay.tsx
@@ -1,11 +1,11 @@
-import sharedClasses from '../styles.module.css';
-import { Votes } from '../types/WebSocket';
-import { IconNotVoted } from './IconNotVoted';
-import { IconObserver } from './IconObserver';
-import { IconVoted } from './IconVoted';
+import sharedClasses from '../../styles.module.css';
+import { Votes } from '../../types/WebSocket';
+import { IconNotVoted } from '../IconNotVoted/IconNotVoted';
+import { IconObserver } from '../IconObserver/IconObserver';
+import { IconVoted } from '../IconVoted/IconVoted';
 import classes from './VotingStateDisplay.module.css';
-import { connectToWebSocket } from './WebSocket';
-import { COLUMN_NAME, COLUMN_VOTED, VOTE_OBSERVER } from '../constants';
+import { connectToWebSocket } from '../WebSocket/WebSocket';
+import { COLUMN_NAME, COLUMN_VOTED, VOTE_OBSERVER } from '../../constants';
 
 const getSortedVotingState = (votes: Votes) => {
   const votedUsers = Object.keys(votes).map((user) => ({
