+++ frontend/src/index-ssr.tsx
@@ -1,5 +1,5 @@
 import renderToString from 'preact-render-to-string';
-import { App } from './Components/App';
+import { App } from './Components/App/App';
 
 export function render() {
   return renderToString(<App />);
