+++ frontend/src/index.tsx
@@ -1,5 +1,5 @@
 import { hydrate } from 'preact';
-import { App } from './Components/App';
+import { App } from './Components/App/App';
 
 // Hydrate instead of "render" is necessary to reuse pre-rendered content
 hydrate(<App />, document.body);
